testing/web-platform/tests/webmessaging/with-ports/026.html

<!doctype html>
<title>object with length as transferable</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function() {
  postMessage('', '*', {length:1});
  onmessage = this.step_func(function(e) {
    assert_array_equals(e.ports, [null]);
    this.done();
  });
});
</script>
